Re-org service code layout
