+++ service/Cargo.toml
@@ -3,10 +3,6 @@ name = "fauxstodian"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]] # Bin to run the fauxstodian gRPC server
-name = "fauxstodian-server"
-path = "src/server.rs"
-
 [dependencies]
 prost = "0.12.1"
 tonic = "0.10.2"
