+++ service/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod proto {
+    tonic::include_proto!("fauxstodian");
+}
+
+pub mod service;
